many fixes for the testsuite; rewrite testrunner
The stdlib tests are now included in the repository. since ubuntu
does not ships them in any package.

The testrunner is completely rewritten. When testrunner.py was originally written
a lot of tests were failing and some of them were hanging. Thus testrunner.py
was focused on proving good reports and could retry the tests with a particular
test case disabled if it detected time out. Now testrunner is somewhat simpler as
we mostly concerned whether the tests pass or not. It also runs the tests
concurrently, to improve the speed at which we get the results back from travis-ci.
Some tests time execution and assert certain timings, this can fail, especially when
the number of workers is bigger than cpu count. The new testrunner will re-try
tests that failed in a sequential run, when nothing else is run. If this fixes
the problem, the test is reported as "succeeded, but only on the second try" and does
reports the whole suite as failed.

.travis.yml is fixed to download binary packages for all dependencies and use system python.

rename test__benchmarks.py to xtest__benchmarks.py to exclude it form testrunner.py.

Exclude test__count from test_thread.py, because it uses function thread._count() we don't implement

Run monkey patched tests twice, first with patch_all(), then with patch_all(Event=True)