resolver_ares: fallback to threadpool
For non-network requests, like "localhost", it seems to be quite hard
to make gevent.ares to do the right (=compatible with stdlib) thing.

So we check if host does not have '.' and if that's the case,
we fallback to using threadpool.

This covers requests like "localhost", "ip6-localhost" and Python-specific
things, like "<broadcast>"

It's possible to disable this by passing use_threadpool=False to ResolverAres
or by settings resolver.pool = None.