Semaphore/Lock: do not do get_hub() in __init__
because threading has a module-level lock(), it means that get_hub() was called by monkey.patch_all()

this is not good, in case use wants to fork off children, but don't actually use gevent in master or if
the user wants to configure the hub.

Also, PySemaphore is now removed (don't really see the point of maintaing the same code twice).