Fixed WindowsError 183 when installing on Windows
When trying to install gevent on Windows, a `WindowsError [183]`
occured:

    Traceback (most recent call last):
      File "util/cythonpp.py", line 801, in <module>
          process_filename(filename, options.output_file)
      File "util/cythonpp.py", line 82, in process_filename
          atomic_write(pyx_filename, py_banner + value)
      File "util/cythonpp.py", line 519, in atomic_write
          os.rename(tmpname, filename)

The reason is, that `atomic_write` writes a `filename.tmp` file first and
then tries to rename the `filename.tmp` file to the real filename.
That fails on Windows, because filename already exists.
Now `atomic_write` writes the `filename.tmp`, then moves `filename`
to `filename.bak` and then renames `filename.tmp` to `filename` and
finally removes `filename.bak`.