Allow subclasses of WSGIHandler to deal with invalid HTTP requests.
Brought up on IRC.

This involves having read_request raise ValueError subclasses instead of directly logging errors and returning False. This method could already raise ValueError when parsing headers, so callers are prepared for this. It shouldn't be a performance issue because (1) the try block was already being established and (2) these are exceptional situations, not the common case.