Add to release notes a reference about the PyPy cpyext memory leak bug.
Also, this comes on the heels of adding back __weakref__ to Semaphore.
This becomes a field in the struct and the field is cleared in
tp_dealloc. It seems like this would trigger the memory leak bug when
taking references to a subclass, but testing seems to show that
doesn't happen. I guess PyPy handles that specially:

>>>> objgraph.show_growth()
>>>> _ = [weakref.ref(gevent.thread.LockType()) for i in range(1000)]; del _
>>>> objgraph.show_growth()
>>>>

[skip ci]