Reuse an existing loop instance when a new hub is created, fixing a crash in a particular use cases. Fixes #237 and fixes #238.
Also clean up usage of hub._threadlocal, eliminating many try/except cases.

The particular crash seems to be in a corner-case usage (generally,
destroying a hub and/or loop seems to be advanced or rare usage), so I
feel pretty safe merging it into a release-track branch. The lifetime
management of the loop may not be totally ideal, but it solves the
issue. loop objects may now live longer (until the death of their
thread) if the hub was destroyed but the loop wasn't.