Handle corecext and corecffi backends in the same way.
This allows building, using and testing corecffi on CPython, which can
make debugging easier. It also simplifies the build steps for PyPy.

Start testing corecffi on CPython too for comparison with PyPy and speed
comparisons with CPython. Currently getting a few unexpected failures
that need investigation; could be bugs in corecffi.