Link to whats new and changelog from PyPI page.
It's a pet peeve of mine when you can't at least directly get to the
changelog for a release from the PyPI page. Before this, it took at
least three clicks and some reading to figure out where they
were.

(Including directly in the page is better, but our notes are kind
of long; maybe we should split them up?)

[skip ci]