Fix #750 by making patched threading.RLock interpret timeout the same as the stdlib.
This required a change to our test_threading_2.py: we weren't testing
patched RLocks at all! Now we run the full set on them.

Also add test_threading.py from Python 3.4 and 3.5 to the test suite to
identify any other differences. A few tests have to be commented out
because they rely on reprs or internal class details, but nothing too
severe or concerning.

Locally tox 3.3, 3.4 and 3.5 are all green.