Fix python bug 13502
http://bugs.python.org/issue13502

A race condition in Event that caused an improper return value if
greenlets were waiting while the event was set, but then immediately
cleared.

This was fixed in Python 3 but not (yet?) backported to Python 2.7, even
though the bug identifies that as a target.

Not only does this behaviour seem more obviously correct to me, it's
more future-proof: user share of Python 2 should be diminishing while
Python 3 share (with the correct behaviour) grows.