Documentation.
Link socket and SSLSocket objects to their version appropriate
documentation on python.org.

Explicitly use the version-specific base socket class for SSLSocket.
This prevents the docs from listing the wrong base (because the docs are
built on Python 3 and so gevent.socket would be gevent._socket3, leading
the base of gevent._ssl2.SSLSocket to be gevent._socket3.socket).

Also, no need to 'from io import BlockingIOError' for Python 3. It
became a builtin in Python 3.3 which is the lowest version we support.