Use subclass of dict for WSGI environ.
Right now, this is only used for suppressing the printing of potentially
sensitive information, but in the future there could be other uses.

This is technically not compliant with PEP3333 which specifies that the
type(environ) must be dict, but it's not clear if that practically
matters anymore (it looks like it might be a holdover from supporting
Python 1.5.2 before one could subclass the builtin dict; see https://mail.python.org/pipermail/web-sig/2003-December/000394.html).

Fixes #779.