Fix #717.
Actually it wasn't quite as dire as described there. Instead, there were
just extra copies of each chunk of data the application yielded.

This commit removes the copies, sending the data directly to the socket.
This should be most noticeable on larger chunks.

Also re-organize the call to close() the app iter closer to where the
app iter gets assigned, making it easier to spot. Add a test for
this---turns out the validator was closing the iterator for us!