Patch existing locks
In projects which dynamically determine whether to activate gevent,
it can be hard not to import a low level module like logging before
gevent. When logging is imported it initialises a threading.RLock
which it uses to protect the logging configuration. If two
greenthreads attempt to claim this lock, the second one will block the
/native/ thread not just itself. As green systems usually only have
one native thread, this will freeze the whole system.

While searching the GC for unsafe RLocks, replace their internal Lock
with a safe one. Or on Python3, where the default RLock implementation
is in C, swap the whole RLock out with a safe Python version.