Fix #800
sslgte279 and ssl3: Move the check for _sslobj to inside the loop
instead of above it. That way if the socket is closed while
reading/writing (actually, while waiting for the read or write event)
then the ValueError will be raised on the next loop iteration (because
closing cancels the read/write events) instead of a TypeError.

This may or may not obviate the chances for an OSError: closed to be
raised.