subprocess: restore backwards-compatibility with python2
In python2, restore_signals doesn't exist, and the default behavior is
to not restore the signals. (This can be done with preexec_fn, if
wanted.) By setting the option to True by default, and not allowing it
to be changed, the behavior becomes mandatory, and is a change from
cpython's behavior on python 2.7.

This change sets the correct default option for py2, and leaves the
py3 version still customizable, maintaining compatibility in both
versions.