Enable the dnspython resolver to work in non-monkey-patched processes
Do this by importing dns specially with sys.modules containing the
gevent modules. This is the technique eventlet uses (but we don't
leave the patched modules sitting around in sys.modules after we're
done; eventlet may have good reason to do that, but this is not a
public API here and is specialized just for dns).