patch out test_ssl.NetworkedBIOTests.test_handshake.
I see in on OS X even without gevent involved. I see it with gevent involved both on the sendall branch and the master branch. I see it on Travis (now) with the sendall patch.
I don't know if it's been seen before the sendall patch on Travis---flaky builds tend to get rerun. None of the other SSL tests are failing, though, and the pypy version of test_ssl.py notes that test_handshake has been removed from CPython altogether.
It looks like it's very dependent on buffering and things like that.