Working on some more socket benchmarks
The good news is that current master is about 10-15% faster for
sendall than 1.2.2 was (e.g., 301ws vs 256ms in Python 3.6). udp
sendto is roughly unaffected (within the margins, based on the native
performance). Moving the chunking implementation of sendall to Cython
doesn't show any improvements (so that's not a bottleneck, at least in
these benchmarks).

The "bad" news is that both UDP and (especially) sendall perform much
worse than native (native does about 47ms for sendall). This is
probably related to the fact that we're doing everything in one
process and one thread, and it is CPU bound; the native process can
use 150% CPU or so, but the gevent version cannot. So the comparison
is not directly meaningful.

[skip ci]