Run queued callbacks when our prepare callback runs.
This solves the timer delay. Add a specific test for this.

We're still sensitive to exactly when these run, for reasons that
aren't clear (comment inline). I want to spend some more time
investigating that.

This disables user-access to prepare watchers on libuv (otherwise
we're right back where we started). It could be worked around if
needed without *too* much trouble.