Make ThreadPool.join respect the loop.min_sleep_time value.
Setting the size was already doing so. By not using the hardcoded value, this prevents a UserWarning from the libuv backend.

Also rename the attribute to  and implement it for all loops.
Add this to the ILoop interface, and test that all loop implementations correctly implement it with zope.interface.verify.verifyObject.

Fixes #1321.