First part of fixing #1349: Allow test.support.is_resource_enabled to do its usual thing in monkey-patched tests (and eventually our own tests).
By default, this will always return true, so nothing will be skipped. An eventual step is to add options to the testrunner to control which resources are available, analogous to what test.regrtest can do.

Because we always spawn fresh subprocesses, we'll have to do this with environment variables and be sure we have the right setup code always invoked.