Make Semaphores fair.
Preserve order of wakeups. Take steps to avoid going back to the O(n^2) old behaviour, though, much like Greenlet does.

This means we auto-unlink links as we call them. Most users won't notice as that's a very low-level method, but baseserver.py was relying on the same rawlink being called every time a semaphore became ready. That was *a lot*, so this eliminates some calls.