Fix #1663 by more gracefully handling non-type arguments.
The documentation specifically says that kill, killall, etc, all take a a type, but
allow passing exception instances.

Also allow passing arbitrary objects; this is also not documented, and not recommended.
The result will be raising a BaseException in the greenlet. That should be better than
silently printing something to stderr, which is what happened before.