Add 3.11 stdlib tests; let the ThreadPool threads exit when idle.
Exiting when idle is because 3.11's test_socket.BasicTCPTest.testDetach() was always failing complaining about a leaked thread. Turns out this thread was from the default threadpool used for DNS resolution.
Something must have changed to cause the first DNS resolution to occur during that test, and hence the first thread to be spawned.  This is opt-in for all thread pools, with the default thread pool opting
in by default; but a config setting allows that to be changed.

Most 3.11 tests pass, except for some in test_subprocess where new functionality was added.