Fix the remainder of the 3.11 subprocess tests; try to stop skipping the tree format tests on mac.
subprocess needed a new check added to check_output, and needed process_group implemented.

because all the leakcheck tests finished successfully after the last commit, and because it sure looks like a refcounting issues in the tree formatting tests (leaked greenlets causing the numbering to be off), try enabling them for this run.

fixes #1872
refs #1867