Monkey-patch queue.LifoQueue to be gevent.queue.LifoQueue.
Works around https://github.com/gevent/gevent/issues/1957

This issue has been opened for almost two years. Until it is resolved, everybody
hitting it has to spend time diagnosing it, implementing the workaround and
deploying it. With this change, we relieve users from having to rediscover,
implement and maintain that kludge and make it easy to remove once the issue is
resolved.