Update to libuv 1.51; fixes #2108.
Windows: Fix build of libev_vfd.h

It's unclear how this was working before as the bad functions, PyInt_AsLong, PyInt_FromLong, do not exist on Python 3.
I think we were getting a backwards compatibility macro from the Cython generated code, so the order of includes relative to
macro definitions must have changed.
This is supported by the fact that Cython 3.1 was released yesterday.

musllinux_1_1 fails to build libuv because of a missing pthread_getname_np. Try with 1_2.